Fixup erroneous use of `cur` and `end` in qmda_queue_add
`cur` and `end` were used to try and send multiple messages to the user. 
However, only one message is ever sent before returning.
This also caused necessary error handling in termsof the rewinding the 
queue not to happen.

We remove this erroneous code and simply supply the user with a 
truncated message in case their buffer is too short.